# node-service/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HM程序员社区 API",
        "version": "1.0.0",
        "description": "HM程序员社区后台服务API文档 - 提供用户管理、内容管理、社区互动等功能",
        "contact": {
            "name": "HM Community Team",
            "email": "admin@hm.com",
            "url": "https://github.com/hm-community"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3001/api",
            "description": "开发环境"
        },
        {
            "url": "https://api.hm-community.com/api",
            "description": "生产环境"
        }
    ],
    "paths": {
...
        "/content": {
            "get": {
                "tags": [
                    "内容管理"
                ],
                "summary": "获取内容列表",
                "description": "分页获取内容列表，支持按类型、分类、标签等条件筛选",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "article",
                                "question",
                                "snippet"
                            ]
                        },
                        "description": "内容类型：article(文章)、question(问答)、snippet(代码片段)"
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "分类ID或标识"
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "标签名称"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "draft",
                                "published",
                                "archived"
                            ]
                        },
                        "description": "内容状态"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1
                        },
                        "description": "页码"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 50,
                            "default": 20
                        },
                        "description": "每页数量"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "latest",
                                "hot",
                                "likes",
                                "views"
                            ]
                        },
                        "description": "排序方式：latest(最新)、hot(热门)、likes(点赞数)、views(浏览数)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "contents": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Content"
                                                    }
                                                },
                                                "pagination": {
                                                    "$ref": "#/components/schemas/Pagination"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "内容管理"
                ],
                "summary": "创建内容",
                "description": "创建新的内容（文章、问答或代码片段）",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "title",
                                    "content",
                                    "type"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "description": "标题",
                                        "minLength": 1,
                                        "maxLength": 200,
                                        "example": "JavaScript异步编程详解"
                                    },
                                    "content": {
                                        "type": "string",
                                        "description": "内容正文（支持Markdown）",
                                        "example": "# JavaScript异步编程\n\n异步编程是JavaScript的重要特性..."
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "article",
                                            "question",
                                            "snippet"
                                        ],
                                        "description": "内容类型",
                                        "example": "article"
                                    },
                                    "categoryId": {
                                        "type": "integer",
                                        "description": "分类ID",
                                        "example": 1
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "description": "标签列表",
                                        "example": [
                                            "JavaScript",
                                            "异步编程",
                                            "Promise"
                                        ]
                                    },
                                    "summary": {
                                        "type": "string",
                                        "description": "摘要",
                                        "maxLength": 500,
                                        "example": "本文详细介绍了JavaScript中的异步编程概念和实践"
                                    },
                                    "coverImage": {
                                        "type": "string",
                                        "description": "封面图片URL"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "内容创建成功"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Content"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "用户ID",
                        "example": 1
                    },
                    "username": {
                        "type": "string",
                        "description": "用户名",
                        "example": "testuser"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "邮箱",
                        "example": "test@example.com"
                    },
                    "nickname": {
                        "type": "string",
                        "description": "昵称",
                        "example": "测试用户"
                    },
                    "avatar": {
                        "type": "string",
                        "description": "头像URL",
                        "example": "/uploads/avatars/default.jpg"
                    },
                    "bio": {
                        "type": "string",
                        "description": "个人简介",
                        "example": "热爱编程的开发者"
                    },
                    "location": {
                        "type": "string",
                        "description": "所在地",
                        "example": "北京"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "user",
                            "admin",
                            "auditor"
                        ],
                        "description": "角色",
                        "example": "user"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "banned"
                        ],
                        "description": "状态",
                        "example": "active"
                    },
                    "level": {
                        "type": "integer",
                        "description": "用户等级",
                        "example": 1
                    },
                    "experience": {
                        "type": "integer",
                        "description": "经验值",
                        "example": 100
                    },
                    "followerCount": {
                        "type": "integer",
                        "description": "粉丝数",
                        "example": 0
                    },
                    "followingCount": {
                        "type": "integer",
                        "description": "关注数",
                        "example": 0
                    },
                    "contentCount": {
                        "type": "integer",
                        "description": "发布内容数",
                        "example": 0
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "注册时间",
                        "example": "2024-01-01T00:00:00.000Z"
                    }
                }
            },
            "Content": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "内容ID",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "description": "标题",
                        "example": "JavaScript异步编程详解"
                    },
                    "summary": {
                        "type": "string",
                        "description": "摘要",
                        "example": "本文详细介绍了JavaScript中的异步编程概念和实践"
                    },
                    "content": {
                        "type": "string",
                        "description": "内容正文",
                        "example": "# JavaScript异步编程\n\n异步编程是JavaScript的重要特性..."
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "article",
                            "question",
                            "snippet"
                        ],
                        "description": "内容类型",
                        "example": "article"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "published",
                            "archived"
                        ],
                        "description": "状态",
                        "example": "published"
                    },
                    "coverImage": {
                        "type": "string",
                        "description": "封面图片URL",
                        "example": "/uploads/covers/default.jpg"
                    },
                    "viewCount": {
                        "type": "integer",
                        "description": "浏览量",
                        "example": 100
                    },
                    "likeCount": {
                        "type": "integer",
                        "description": "点赞数",
                        "example": 10
                    },
                    "favoriteCount": {
                        "type": "integer",
                        "description": "收藏数",
                        "example": 5
                    },
                    "commentCount": {
                        "type": "integer",
                        "description": "评论数",
                        "example": 3
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "创建时间",
                        "example": "2024-01-01T00:00:00.000Z"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "更新时间",
                        "example": "2024-01-01T00:00:00.000Z"
                    },
                    "author": {
                        "$ref": "#/components/schemas/User"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    }
                }
            },
...
            "Category": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "分类ID",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "description": "分类名称",
                        "example": "前端开发"
                    },
                    "slug": {
                        "type": "string",
                        "description": "分类标识",
                        "example": "frontend"
                    },
                    "description": {
                        "type": "string",
                        "description": "分类描述",
                        "example": "前端技术相关内容"
                    },
                    "icon": {
                        "type": "string",
                        "description": "分类图标",
                        "example": "🎨"
                    },
                    "contentCount": {
                        "type": "integer",
                        "description": "内容数量",
                        "example": 50
                    }
                }
            },
            "Tag": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "标签ID",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "description": "标签名称",
                        "example": "JavaScript"
                    },
                    "slug": {
                        "type": "string",
                        "description": "标签标识",
                        "example": "javascript"
                    },
                    "color": {
                        "type": "string",
                        "description": "标签颜色",
                        "example": "#f7df1e"
                    },
                    "useCount": {
                        "type": "integer",
                        "description": "使用次数",
                        "example": 100
                    }
                }
            },
            "Pagination": {
                "type": "object",
                "properties": {
                    "current": {
                        "type": "integer",
                        "description": "当前页码",
                        "example": 1
                    },
                    "total": {
                        "type": "integer",
                        "description": "总页数",
                        "example": 10
                    },
                    "count": {
                        "type": "integer",
                        "description": "总记录数",
                        "example": 200
                    },
                    "per_page": {
                        "type": "integer",
                        "description": "每页记录数",
                        "example": 20
                    }
                }
            },
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "操作失败"
                    },
                    "code": {
                        "type": "string",
                        "description": "错误代码",
                        "example": "VALIDATION_ERROR"
                    },
                    "errors": {
                        "type": "array",
                        "description": "详细错误信息",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "responses": {
            "BadRequest": {
                "description": "请求参数错误",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "未授权访问",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "请在请求头中添加 Authorization: Bearer {token}"
            }
        }
    },
...
}